requirements.txt
asgiref==3.2.10
blis==0.4.1
catalogue==1.0.0
certifi==2020.11.8
chardet==3.0.4
click==7.1.2
cycler==0.10.0
cymem==2.0.4
decorator==4.4.2
Django==3.1.2
djangorestframework==3.12.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
idna==2.10
joblib==0.17.0
kiwisolver==1.3.1
matplotlib==3.3.3
murmurhash==1.0.4
networkx==2.5
nltk==3.5
numpy==1.19.4
pandas==1.1.4
Pillow==8.0.1
plac==1.1.3
preshed==3.0.4
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
regex==2020.11.13
requests==2.25.0
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
spacy==2.3.2
sqlparse==0.4.1
srsly==1.0.4
thinc==7.4.1
threadpoolctl==2.1.0
tqdm==4.53.0
urllib3==1.26.2
wasabi==0.8.0
